Update schema script
